requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
alembic==1.12.0
async-timeout==4.0.3
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.3.0
discord.py==2.3.2
frozenlist==1.4.0
greenlet==3.0.0
idna==3.4
Mako==1.2.4
MarkupSafe==2.1.3
multidict==6.0.4
python-dotenv==1.0.0
requests==2.31.0
SQLAlchemy==2.0.21
typing_extensions==4.8.0
urllib3==2.0.6
yarl==1.9.2
